requirements.txt
# Core dependencies
google-generativeai>=0.8.3
langchain
langchain-google-genai
openai
langchain-openai

# Data validation and structuring
pydantic

# Ghost CMS API communication
requests
PyJWT

# Google API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# HTML/Markdown processing
markdown
markdownify
beautifulsoup4

# Environment variable management
python-dotenv

# Social Media APIs
Mastodon.py

# Image Generation & Frontmatter
python-frontmatter
Pillow